Add web invite entry points
